requirements.txt
# --- PDF ---
PyMuPDF>=1.23.0
pdfplumber==0.11.0
Pillow==10.1.0

# --- OCR ---
pytesseract==0.3.13
opencv-python==4.8.1.78
# Paddle
# paddlepaddle==2.6.0
# paddleocr==2.7.0.3

# --- LLM / Embeddings ---
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
huggingface-hub==0.20.2
langchain==0.1.0
langchain-community==0.0.10

# --- Vector DB ---
pinecone-client==3.0.0

# --- Utils ---
numpy==1.26.2
pandas==2.1.4
pydantic==2.5.3
PyYAML==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0
